@@ -1,3 +1,3 @@
-module Types::BaseInterface
+module Types::Base::Interface
   include GraphQL::Schema::Interface
 end
