@@ -1,9 +1,9 @@
-class InfluxUtil
-  class << self
-
-    def has_influx?
-      File.exist?("/etc/influxdb/influxdb.conf")
-    end
-
-  end
-end
+# class InfluxUtil
+#   class << self
+#
+#     def has_influx?
+#       File.exist?("/etc/influxdb/influxdb.conf")
+#     end
+#
+#   end
+# end
