@@ -3,9 +3,7 @@ Rails.application.routes.draw do
   devise_for :users
 
   # ----- GRAPHQL -----
-  if Rails.env.development?
-    mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/graphql"
-  end
+  mount GraphiQL::Rails::Engine, at: "/graphiql", graphql_path: "/graphql"
   post "/graphql", to: "graphql#execute"
 
   # ----- INFO ROUTES -----
